Auto stash before merge of "Solenn" and "origin/main"

+++ tp-vol/tp-vol-dao/src/main/java/sopra/vol/repository/jpa/ReservationRepositoryJpa.java
@@ -47,7 +47,7 @@ public class ReservationRepoJpa implements IReservationRepositoryJpa{
 		
 	}
 	
-	public Reservation findById(Integer numero) {
+	public Reservation findById(Long numero) {
 		Reservation reservation = null;
 
 		EntityManager em = null;
